feat: sécurité du site

+++ src/Controller/AvionController.php
@@ -17,14 +17,21 @@ class AvionController extends AbstractController
     #[Route('/', name: 'app_avion_index', methods: ['GET'])]
     public function index(AvionRepository $avionRepository): Response
     {
-        return $this->render('avion/index.html.twig', [
-            'avions' => $avionRepository->findAll(),
-        ]);
+        if ($this->isGranted('ROLE_ADMIN')) {
+            return $this->render('avion/index.html.twig', [
+                'avions' => $avionRepository->findAll(),
+            ]);
+        }else{
+            return $this->redirectToRoute('app_home');
+        }
     }
 
     #[Route('/new', name: 'app_avion_new', methods: ['GET', 'POST'])]
     public function new(Request $request, EntityManagerInterface $entityManager): Response
     {
+        if (!$this->isGranted('ROLE_ADMIN')) {
+            return $this->redirectToRoute('app_home');
+        }
         $avion = new Avion();
         $form = $this->createForm(AvionType::class, $avion);
         $form->handleRequest($request);
@@ -45,6 +52,9 @@ class AvionController extends AbstractController
     #[Route('/{id}', name: 'app_avion_show', methods: ['GET'])]
     public function show(Avion $avion): Response
     {
+        if (!$this->isGranted('ROLE_ADMIN')) {
+            return $this->redirectToRoute('app_home');
+        }
         return $this->render('avion/show.html.twig', [
             'avion' => $avion,
         ]);
@@ -53,6 +63,9 @@ class AvionController extends AbstractController
     #[Route('/{id}/edit', name: 'app_avion_edit', methods: ['GET', 'POST'])]
     public function edit(Request $request, Avion $avion, EntityManagerInterface $entityManager): Response
     {
+        if (!$this->isGranted('ROLE_ADMIN')) {
+            return $this->redirectToRoute('app_home');
+        }
         $form = $this->createForm(AvionType::class, $avion);
         $form->handleRequest($request);
 
@@ -71,6 +84,9 @@ class AvionController extends AbstractController
     #[Route('/{id}', name: 'app_avion_delete', methods: ['POST'])]
     public function delete(Request $request, Avion $avion, EntityManagerInterface $entityManager): Response
     {
+        if (!$this->isGranted('ROLE_ADMIN')) {
+            return $this->redirectToRoute('app_home');
+        }
         if ($this->isCsrfTokenValid('delete'.$avion->getId(), $request->request->get('_token'))) {
             $entityManager->remove($avion);
             $entityManager->flush();
